refactor: modularize order service
